# documentation/dds-mano-g06-2.0-swagger.yaml
openapi: 3.0.0
info:
  version: '2.0'
  title: DDS TP Carbono TPA - G06
  description: Gestión de Organizaciones y el cálculo de la Huella de Carbono
servers:
  - url: https://api.ddsg06.com #cambiar
paths:
...
  /api/organizacion/{id}:
    get:
      summary: Organización
      description: Mostrar la organización indicada
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
    put:
      summary: Organización
      description: Modificar la organización indicada
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
    delete:
      summary: Organización
      description: Borrar la organización indicada
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
  /api/organizacion/{id}/batch:
    get:
      summary: Batches de Mediciones
      description: Mostrar los batches de mediciones de la organización indicada
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
    post:
      summary: Batch de Mediciones
      description: Cargar un batch de mediciones para la organización indicada
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
...
  /api/organizacion/{id}/medicion:
    get:
      summary: Mediciones
      description: Mostrar las mediciones de la organización indicada
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/OrgError'
...
  /api/organizacion/{id}/reporte:
    get:
      summary: Reporte
      description: Obtener el reporte de la organización indicada
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/OrgError'
...
  /api/miembro/{id}:
    put:
      summary: Miembro
      description: Modificar el miembro indicado
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/MiembroError'
    delete:
      summary: Miembro
      description: Eliminar el miembro indicado
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/MiembroError'
...
  /api/agente/{id}:
    get:
      summary: Agente
      description: Obtener el agente sectorial indicado
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/AgenteError'
    put:
      summary: Agente
      description: Modificar el agente sectorial indicado
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/AgenteError'
    delete:
      summary: Agente
      description: Eliminar el agente sectorial indicado
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/AgenteError'
  /api/agente/{id}/reporte:
    get:
      summary: Reporte
      description: Obtener el reporte del agente sectorial indicado
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/AgenteError'
...
  /api/trayecto/{id}:
    delete:
      summary: Trayecto
      description: Eliminar del sistema el trayecto indicado
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - API
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/TrayectoError'
...
  /miembro/{id}/trayecto:
    get:
      summary: Trayectos
      description: Mostrar los trayectos del miembro indicado
      parameters: 
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
        - in: query
          name: action
          schema:
            $ref: '#/components/schemas/AccionTrayectos'
      tags:
        - WEB
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/MiembroError'
    post:
      summary: Cargar Trayecto
      description: Cargar un trayecto para el miembro indicado
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/Id'
          required: true
        - in: query
          name: compartido
          schema:
            $ref: '#/components/schemas/Boolean'
          required: true
        - in: query
          name: fecha
          schema: 
            $ref: '#/components/schemas/Fecha'
          required: true
        - in: query
          name: transporte
          schema:
            $ref: '#/components/schemas/Id'
          required: true
        - in: query
          name: parada
          schema:
            $ref: '#/components/schemas/Id'
          required: false
        - in: query
          name: lat-inicial
          schema:
            $ref: '#/components/schemas/Coor'
          required: false
        - in: query
          name: lon-inicial
          schema:
            $ref: '#/components/schemas/Coor'
          required: false
        - in: query
          name: lat-final
          schema:
            $ref: '#/components/schemas/Coor'  
          required: false
        - in: query
          name: lon-final
          schema:
            $ref: '#/components/schemas/Coor'
          required: false
        - in: query
          name: transporte-nuevo
          schema:
            $ref: '#/components/schemas/Id'
          required: false
        - in: query
          name: lat-inicial-nueva
          schema:
            $ref: '#/components/schemas/Coor'
          required: false
        - in: query
          name: lon-inicial-nueva
          schema:
            $ref: '#/components/schemas/Coor'
          required: false
        - in: query
          name: lat-final-nueva
          schema:
            $ref: '#/components/schemas/Coor'
          required: false
        - in: query
          name: lon-final-nueva
          schema:
            $ref: '#/components/schemas/Coor'
          required: false
        #FALTAN LAS DIRECCIONES!
      tags:
        - WEB
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/MiembroError'
...
  /miembro/{miembroId}/trayecto/{trayectoId}:
    get:
      summary: Trayecto
      description: Mostrar el trayecto indicado del miembro indicado
      parameters: 
        - in: path
          name: miembroId
          schema:
            $ref: '#/components/schemas/Id'
          required: true
        - in: path
          name: trayectoId
          schema:
            $ref: '#/components/schemas/Id'
          required: true
        - in: query
          name: action
          schema:
            $ref: '#/components/schemas/AccionTrayectos'
      tags:
        - WEB
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/MiembroTrayectoError'
    post:
      summary: Modificar Trayecto
      description: Modificar el trayecto indicado del miembro indicado
      parameters: 
        - in: path
          name: miembroId
          schema:
            $ref: '#/components/schemas/Id'
          required: true
        - in: path
          name: trayectoId
          schema:
            $ref: '#/components/schemas/Id'
          required: true
        #PONER LOS QUERY STRINGS
      tags:
        - WEB
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/MiembroTrayectoError'
    delete:
      summary: Borrar Trayecto
      description: Borrar el trayecto indicado del miembro indicado (no se elimina del sistema)
      parameters: 
        - in: path
          name: miembroId
          schema:
            $ref: '#/components/schemas/Id'
          required: true
        - in: path
          name: trayectoId
          schema:
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - WEB
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/MiembroTrayectoError'
...
  /organizacion/{id}/reporte:
    get:
      summary: Reporte
      description: Mostrar reporte generado para la organización indicada
      parameters: 
        - in: path
          name: id
          schema: 
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - WEB
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/OrgError'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/NoAutorizado'   
    post:
      summary: Generar Reporte
      description: Generar el reporte para la organización indicada
      parameters: 
        - in: path
          name: id
          schema: 
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - WEB
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/OrgError'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/NoAutorizado'
  /agente/{id}/reporte:
    get:
      summary: Reporte
      description: Mostrar reporte generado para el agente indicado
      parameters: 
        - in: path
          name: id
          schema: 
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - WEB
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/OrgError'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/NoAutorizado'   
    post:
      summary: Generar Reporte
      description: Generar el reporte para el agente indicado
      parameters: 
        - in: path
          name: id
          schema: 
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - WEB
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/OrgError'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/NoAutorizado'
  /agente/{id}/organizacion:
    get:
      summary: Organizaciones
      description: Mostrar las organizaciones del agente indicado
      parameters: 
        - in: path
          name: id
          schema: 
            $ref: '#/components/schemas/Id'
          required: true
      tags:
        - WEB
      responses:
        '200':
          $ref: '#/components/responses/Exito'
        '400':
          $ref: '#/components/responses/OrgError'
        '401':
          $ref: '#/components/responses/NoAutenticado'
        '403':
          $ref: '#/components/responses/NoAutorizado'  
components:
  responses:
    Exito:
      description: Response exitosa
    OrgError:
      description: Organizacion de id inexistente
...
    TrayectoError:
      description: Trayecto de id inexistente
    NoAutorizado:
      description: Acceso sin autorización
    NoAutenticado:
      description: Acceso sin usuario logueado
...
    MiembroError:
      description: Miembro de id inexistente
    AgenteError:
      description: Agente Sectorial de id inexistente
    MiembroTrayectoError:
      description: Miembro o Trayecto de id inexistente
  schemas:
    Id:
      type: integer
      minimum: 0
    Coor:
      type: number
    Fecha:
      type: string
      description: mes/año
...
    AccionTrayectos:
      type: string
      enum: [create, list, view]
    Boolean:
      type: boolean
...